Change `angular.service` example to lowerCamelCase
As is indicated just a couple of paragraphs above this example, we should use "lowerCamelCase for all other services" that don't return constructor functions. In this example since the constructor is being registered via `angular.service`, it will return an already-constructed instance, not a constructor function, when injected. So the name of the service should be written lowerCamelCase.